+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Unreleased
+
+##### Enhancements:
+* Renamed the `deployment.environment` resource attribute to `deployment.environment.name`.
+
 ### Version 2.3.0 - 2026-04-22
 
 ##### New features: